requirements.txt
# main
langchain==0.3.26
langchain-core<1.0.0
langgraph<0.3.0

# tool
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-openai==0.3.28

# util
typing-extensions
python-dotenv
fastapi
uvicorn
jinja2

# document
sentence-transformers==5.0.0
transformers==4.53.3
chromadb
PyMuPDF
PyPDF2
reportlab
